+++ R/define_water.R
@@ -2,13 +2,13 @@
 #'
 #' @description This function takes user-defined water quality parameters and creates an S4 "water" class object that forms the input and output of all tidywater models.
 #'
-#' @details Carbonate balance is calculated and units are converted to mol/L. Ionic strength is determined from ions, TDS, or conductivity. Missing values are handled by defaulting to 0 or
-#' NA. Calcium hardness defaults to 65% of the total hardness because that falls within a typical range. For best results
-#' manually specify all ions in the define_water arguments. The following equations are used to determine ionic strength:
-#' Ionic strength (if TDS provided): Crittenden et al. (2012) equation 5-38
-#' Ionic strength (if electrical conductivity provided): Snoeyink & Jenkins (1980)
-#' Ionic strength (from ion concentrations): Lewis and Randall (1921), Crittenden et al. (2012) equation 5-37
-#' Temperature correction of dielectric constant (relative permittivity): Harned and Owen (1958), Crittenden et al. (2012) equation 5-45.
+#' @details Carbonate balance is calculated and units are converted to mol/L. Ionic strength is determined from ions, TDS, or conductivity.
+#' Missing values are handled by defaulting to 0 or NA.
+#' Calcium defaults to 65 percent of the total hardness when not specified. DOC defaults to 95 percent of TOC.
+#' @source Crittenden et al. (2012) equation 5-38 - ionic strength from TDS
+#' @source Snoeyink & Jenkins (1980) - ionic strength from conductivity
+#' @source Lewis and Randall (1921), Crittenden et al. (2012) equation 5-37 - ionic strength from ion concentrations
+#' @source Harned and Owen (1958), Crittenden et al. (2012) equation 5-45 - Temperature correction of dielectric constant (relative permittivity)
 #'
 #' @param ph water pH
 #' @param temp Temperature in degree C
@@ -137,6 +137,10 @@ define_water <- function(ph, temp = 25, alk, tot_hard, ca, mg, na, k, cl, so4,
     warning("Missing value for DOC. Default value of 95% of TOC will be used.")
     doc <- toc * 0.95
     estimated <- paste(estimated, "doc", sep = "_")
+  }
+
+  if (tot_nh3 > 0 & (free_chlorine > 0 | combined_chlorine > 0)) {
+    warning("Both chlorine and ammonia are present and may form chloramines.\nUse chemdose_chloramine for breakpoint caclulations.")
   }
 
   uv254 <- ifelse(missing(uv254), NA_real_, uv254)