+++ R/general.R
@@ -54,8 +54,8 @@ summarize_wq <- function(water, params = c("general")) {
     ))
 
   gen_tab <- knitr::kable(general,
-    format = "simple",
-    col.names = c("General water quality parameters", "Result", "Units")
+                          format = "simple",
+                          col.names = c("General water quality parameters", "Result", "Units")
   )
 
   # Compile major ions
@@ -74,10 +74,10 @@ summarize_wq <- function(water, params = c("general")) {
     pivot_longer(c(Na:CO3), names_to = "ion", values_to = "c_mg")
 
   ions_tab <- knitr::kable(ions,
-    format = "simple",
-    col.names = c("Major ions", "Concentration (mg/L)"),
-    # format.args = list(scientific = TRUE),
-    digits = 2
+                           format = "simple",
+                           col.names = c("Major ions", "Concentration (mg/L)"),
+                           # format.args = list(scientific = TRUE),
+                           digits = 2
   )
 
   # Compile corrosion indices
@@ -99,8 +99,8 @@ summarize_wq <- function(water, params = c("general")) {
     )
 
   corr_tab <- knitr::kable(corrosion,
-    format = "simple",
-    col.names = c("Corrosion Indices", "Result", "Units", "Recommended")
+                           format = "simple",
+                           col.names = c("Corrosion Indices", "Result", "Units", "Recommended")
   )
 
   # Compile DBPs
@@ -137,13 +137,13 @@ summarize_wq <- function(water, params = c("general")) {
     mutate(result = round(result, 2))
 
   thm_tab <- knitr::kable(tthm,
-    format = "simple",
-    col.names = c("THMs", "Modeled concentration (ug/L)")
+                          format = "simple",
+                          col.names = c("THMs", "Modeled concentration (ug/L)")
   )
 
   haa_tab <- knitr::kable(haa5,
-    format = "simple",
-    col.names = c("HAAs", "Modeled concentration (ug/L)")
+                          format = "simple",
+                          col.names = c("HAAs", "Modeled concentration (ug/L)")
   )
 
   # Print tables
@@ -177,15 +177,15 @@ summarise_wq <- summarize_wq
 #' @import ggplot2
 #'
 #' @examples
-#' water_defined <- define_water(7, 20, 50, 100, 80, 10, 10, 10, 10, tot_po4 = 1)
-#' plot_ions(water_defined)
+#' water<- define_water(7, 20, 50, 100, 20, 10, 10, 10, 10, tot_po4 = 1)
+#' plot_ions(water)
 #'
 #' @export
 #'
 #' @returns A ggplot object displaying the water's ion balance.
 #'
 plot_ions <- function(water) {
-  type <- concentration <- label_pos <- ion <- label_y <- Na <- OH <- NULL # Quiet RCMD check global variable note
+  type <- concentration <- label_pos <- ion <- label_y <- label <- repel_label <- Na <- OH <- NULL # Quiet RCMD check global variable note
   if (!methods::is(water, "water")) {
     stop("Input water must be of class 'water'. Create a water using define_water.")
   }
@@ -209,45 +209,42 @@ plot_ions <- function(water) {
     OH = water@oh
   )
 
-  ions %>%
+  plot <- ions %>%
     tidyr::pivot_longer(c(Na:OH), names_to = "ion", values_to = "concentration") %>%
-    dplyr::mutate(type = case_when(ion %in% c("Na", "Ca", "Mg", "K", "NH4", "H") ~ "Cations", TRUE ~ "Anions")) %>%
-    dplyr::arrange(type, concentration) %>%
-    dplyr::mutate(
-      label_pos = cumsum(concentration) - concentration / 2, .by = type,
-      label_y = case_when(type == "Cations" ~ 2 - .2, TRUE ~ 1 - .2)
+    dplyr::mutate(type = case_when(ion %in% c("Na", "Ca", "Mg", "K", "NH4", "H") ~ "Cations", TRUE ~ "Anions"),
+                  ion = factor(ion, levels = c("Ca", "Mg", "Na", "K", "NH4", "H",
+                                               "HCO3", "CO3", "SO4", "Cl", "H2PO4", "HPO4", "PO4", "OCl", "OH")),
+                  concentration = case_when(is.na(concentration) ~ 0, TRUE ~ concentration)) %>%
+    dplyr::arrange(ion) %>%
+    dplyr::mutate(label_pos = cumsum(concentration) - concentration / 2, .by = type,
+                  label_y = case_when(type == "Cations" ~ 2 - .2, TRUE ~ 1 - .2)
     ) %>%
-    dplyr::filter(!is.na(concentration)) %>%
-    ggplot(aes(x = concentration, y = type, fill = stats::reorder(ion, -concentration))) +
-    geom_bar(
-      stat = "identity",
-      width = 0.5,
-      alpha = 0.5,
-      color = "black"
-    ) +
-    geom_text(aes(x = label_pos, label = ifelse(concentration > 10e-5, ion, ""), fontface = "bold", angle = 90),
-      size = 3.5
-    ) +
+    dplyr::filter(!is.na(concentration),
+                  concentration > 0) %>%
+    dplyr::mutate(label = case_when(concentration > 10e-5 ~ ion, TRUE ~ ""),
+                  repel_label = case_when(concentration <= 10e-5 & concentration > 10e-7 ~ ion, TRUE ~ "")) %>%
+    dplyr::mutate(ion = forcats::fct_rev(ion))
+
+  plot %>%
+    ggplot(aes(x = concentration, y = type, fill = ion)) +
+    geom_bar(stat = "identity", width = 0.5, alpha = 0.5, color = "black") +
+    geom_text(aes(label = label, fontface = "bold", angle = 90),
+              size = 3.5, position = position_stack(vjust = 0.5)) +
     ggrepel::geom_text_repel(
       aes(
         x = label_pos, y = label_y,
-        label = ifelse(concentration <= 10e-5 & concentration > 0, ion, ""),
+        label = repel_label,
         fontface = "bold"
       ),
       size = 3.5,
       nudge_y = -.2,
-      seed = 555
-    ) +
+      seed = 555) +
     theme_bw() +
-    theme(axis.title = element_text(face = "bold")) +
-    labs(
-      x = "Concentration (eq/L)",
-      y = "Major Cations and Anions",
-      subtitle = paste0("pH=", water@ph, "\nAlkalinity=", water@alk)
-    ) +
-    guides(fill = "none")
+    theme(axis.title = element_text(face = "bold"),
+          legend.position = "none") +
+    labs(x = "Concentration (eq/L)", y = "Major Cations and Anions",
+         subtitle = paste0("pH=", water@ph, "\nAlkalinity=", water@alk))
 }
-
 
 #' @title Calculate unit conversions for common compounds
 #'
@@ -288,25 +285,25 @@ convert_units <- function(value, formula, startunit = "mg/L", endunit = "M") {
   # Determine multiplier for order of magnitude conversion
   # In the same list, no multiplier needed
   if ((startunit %in% milli_list & endunit %in% milli_list) |
-    (startunit %in% stand_list & endunit %in% stand_list) |
-    (startunit %in% nano_list & endunit %in% nano_list) |
-    (startunit %in% mcro_list & endunit %in% mcro_list)) {
+      (startunit %in% stand_list & endunit %in% stand_list) |
+      (startunit %in% nano_list & endunit %in% nano_list) |
+      (startunit %in% mcro_list & endunit %in% mcro_list)) {
     multiplier <- 1
     # m - standard, n-u, u-n
   } else if ((startunit %in% milli_list & endunit %in% stand_list) |
-    (startunit %in% mcro_list & endunit %in% milli_list) |
-    (startunit %in% nano_list & endunit %in% mcro_list)) {
+             (startunit %in% mcro_list & endunit %in% milli_list) |
+             (startunit %in% nano_list & endunit %in% mcro_list)) {
     multiplier <- 1e-3
   } else if ((startunit %in% stand_list & endunit %in% milli_list) |
-    (startunit %in% milli_list & endunit %in% mcro_list) |
-    (startunit %in% mcro_list & endunit %in% nano_list)) {
+             (startunit %in% milli_list & endunit %in% mcro_list) |
+             (startunit %in% mcro_list & endunit %in% nano_list)) {
     multiplier <- 1e3
     # u - standard
   } else if ((startunit %in% mcro_list & endunit %in% stand_list) |
-    (startunit %in% nano_list & endunit %in% milli_list)) {
+             (startunit %in% nano_list & endunit %in% milli_list)) {
     multiplier <- 1e-6
   } else if ((startunit %in% stand_list & endunit %in% mcro_list) |
-    (startunit %in% milli_list & endunit %in% nano_list)) {
+             (startunit %in% milli_list & endunit %in% nano_list)) {
     multiplier <- 1e6
     # n - standard
   } else if (startunit %in% nano_list & endunit %in% stand_list) {
@@ -324,10 +321,10 @@ convert_units <- function(value, formula, startunit = "mg/L", endunit = "M") {
   # Determine relevant molar weight
   if (formula %in% colnames(tidywater::mweights)) {
     if ((startunit %in% caco_list & endunit %in% c(mole_list, eqvl_list)) |
-      (endunit %in% caco_list & startunit %in% c(mole_list, eqvl_list))) {
+        (endunit %in% caco_list & startunit %in% c(mole_list, eqvl_list))) {
       molar_weight <- caco3_mw
     } else if ((startunit %in% n_list & endunit %in% c(mole_list, eqvl_list)) |
-      (endunit %in% n_list & startunit %in% c(mole_list, eqvl_list))) {
+               (endunit %in% n_list & startunit %in% c(mole_list, eqvl_list))) {
       molar_weight <- n_mw
     } else {
       molar_weight <- as.numeric(tidywater::mweights[formula])
@@ -380,8 +377,8 @@ convert_units <- function(value, formula, startunit = "mg/L", endunit = "M") {
     value / molar_weight * n_mw
     # same lists
   } else if ((startunit %in% gram_list & endunit %in% gram_list) |
-    (startunit %in% mole_list & endunit %in% mole_list) |
-    (startunit %in% eqvl_list & endunit %in% eqvl_list)) {
+             (startunit %in% mole_list & endunit %in% mole_list) |
+             (startunit %in% eqvl_list & endunit %in% eqvl_list)) {
     value * multiplier
   } else {
     stop("Units not supported")
@@ -470,44 +467,72 @@ validate_water <- function(water, slots) {
   }
 }
 
-construct_helper <- function(df, num_arguments, str_arguments) {
-  all_arguments <- c(names(num_arguments), names(str_arguments))
+validate_water_helpers <- function(df, input_water) {
+  # Make sure input_water column is in the dataframe and is a water class.
 
-  inputs_arg <- do.call(expand.grid, num_arguments) %>%
-    select_if(~ any(. != 0))
-
-  if (any(sapply(str_arguments, length) > 1)) {
-    inputs_arg <- inputs_arg %>%
-      cross_join(do.call(expand.grid, str_arguments))
+  if (!(input_water %in% colnames(df))) {
+    stop("Specified input_water column not found. Check spelling or create a water class column using define_water_chain().")
   }
-
-  inputs_col <- df %>%
-    subset(select = names(df) %in% all_arguments) %>%
-    # add row number for joining
-    mutate(ID = row_number())
-
-  if (any(all_arguments %in% colnames(inputs_arg) & all_arguments %in% colnames(inputs_col))) {
-    stop("Argument was applied as both a function argument and a data frame column. Choose one input method.")
+  if (!all(sapply(df[[input_water]], function(x) methods::is(x, "water")))) {
+    stop("Specified input_water does not contain water class objects. Use define_water_chain() or specify a different column.")
   }
+}
 
-  arguments <- inputs_col %>%
-    cross_join(inputs_arg)
-
-  if (!all(all_arguments %in% colnames(arguments))) {
-    if (!all(names(num_arguments) %in% colnames(arguments))) {
-      warning("Numeric arguments missing or set to 0. Add them as a column or function argument.")
+validate_args <- function(num_args = list(), str_args = list(), log_args = list(), misc_args = list()) {
+  all_args <- c(num_args, str_args, log_args, misc_args)
+  for (arg in names(all_args)) {
+    if (is.null(all_args[[arg]])) {
+      stop("argument '", arg, "' is missing, with no default")
     }
+  }
+  for (arg in names(num_args)) {
+    if (!is.numeric(num_args[[arg]])) {
+      stop("argument '", arg, "' must be numeric.")
+    }
+  }
+  for (arg in names(str_args)) {
+    if (!is.character(str_args[[arg]])) {
+      stop("argument '", arg, "' must be specified as a string.")
+    }
+  }
+  for (arg in names(log_args)) {
+    if (!is.logical(log_args[[arg]])) {
+      stop("argument '", arg, "' must be either TRUE or FALSE.")
+    }
+  }
+}
 
-    missing_args <- do.call(expand.grid, num_arguments) %>%
-      cross_join(do.call(expand.grid, str_arguments)) %>%
-      subset(select = !names(.) %in% names(arguments)) %>%
-      unique()
+construct_helper <- function(df, all_args) {
+  # Get the names of each argument type
+  all_arguments <- names(all_args)
+  from_df <- names(all_args[all_args == "use_col"])
 
-    arguments <- arguments %>%
-      cross_join(missing_args)
+  from_new <- all_args[all_args != "use_col"]
+  if (length(from_new) > 0) {
+    from_columns <- from_new[sapply(from_new, function(x) any(inherits(x, "quosure")))]
+  } else {
+    from_columns <- list()
   }
 
-  return(arguments)
+  from_inputs <- setdiff(from_new, from_columns)
+
+  inputs_arg <- do.call(expand.grid, list(from_inputs, stringsAsFactors = FALSE))
+
+
+  if (any(colnames(df) %in% colnames(inputs_arg))) {
+    stop("Argument was applied as a function argument, but the column already exists in the data frame. Remove argument or rename dataframe column.")
+  }
+
+  # Get the new names for relevant columns
+  final_names <- stats::setNames(as.list(all_arguments), all_arguments)
+  for (arg in names(from_columns)) {
+    final_names[[arg]] <- rlang::as_name(from_columns[[arg]])
+  }
+
+  return(list(
+    "new_cols" = as.list(inputs_arg),
+    "final_names" = as.list(final_names)
+  ))
 }
 
 
@@ -584,8 +609,8 @@ K_temp_adjust <- function(deltah, ka, temp) {
 calculate_ionicstrength <- function(water) {
   # From all ions: IS = 0.5 * sum(M * z^2)
   0.5 * (sum(water@na, water@cl, water@k, water@hco3, water@h2po4, water@h, water@oh, water@ocl,
-    water@f, water@br, water@bro3, water@nh4,
-    na.rm = TRUE
+             water@f, water@br, water@bro3, water@nh4,
+             na.rm = TRUE
   ) * 1^2 +
     sum(water@ca, water@mg, water@so4, water@co3, water@hpo4, water@mn, na.rm = TRUE) * 2^2 +
     sum(water@po4, water@fe, water@al, na.rm = TRUE) * 3^2)