+++ tests/testthat/test-blend_waters.R
@@ -68,6 +68,18 @@ test_that("Blend waters warns when some slots are NA.", {
   expect_warning(blend_waters(c(water1, water2), c(.5, .5)), "ca.+na")
 })
 
+test_that("Blend waters warns about chloramines.", {
+  water1 <- suppressWarnings(define_water(7, 20, 50, free_chlorine = 2))
+  water2 <- suppressWarnings(define_water(7.5, 20, 100, tot_nh3 = 2))
+
+  water3 <- suppressWarnings(define_water(8, 22, 13, combined_chlorine = 3))
+  water4 <- suppressWarnings(define_water(8, 22, 13) %>% chemdose_ph(nh4oh = 30))
+
+  expect_warning(blend_waters(c(water1, water2), c(.5, .5)), "breakpoint+")
+  expect_warning(blend_waters(c(water3, water4), c(.25, .75)), "breakpoint+")
+  expect_no_warning(blend_waters(c(water1, water3), c(.20, .80)))
+})
+
 ################################################################################*
 ################################################################################*
 # blend_waters helpers ----
@@ -141,6 +153,22 @@ test_that("blend_waters_once can handle different ways to input ratios", {
   expect_equal(blend2$ph, blend3$ph)
 })
 
+
+test_that("blend_waters_once can handle water columns mixed with objects", {
+
+  water4 <- water_df %>%
+    slice(1:3) %>%
+    define_water_chain("A")
+  water5 <- define_water(ph = 7.9, temp = 20, alk = 50, tot_hard = 50, ca = 13, mg = 4, na = 20, k = 20,
+                         cl = 30, so4 = 20, tds = 200, cond = 100, toc = 2, doc = 1.8, uv254 = 0.05)
+
+  blend4 <- water4 %>%
+    blend_waters_once(c("A", water5), c(.5,.5))
+
+  expect_equal(water_df$ph[1], blend4$ph[1])
+
+})
+
 # Test that blend_waters_chain outputs are the same as base function, blend_waters
 test_that("blend_waters_chain outputs are the same as base function, blend_waters", {
   water1 <- suppressWarnings(define_water(
@@ -195,7 +223,7 @@ test_that("blend_waters_chain can handle different ways to input ratios", {
     chemdose_ph_chain(input_water = "balanced_water", naoh = 20) %>%
     blend_waters_chain(waters = c("balanced_water", "dosed_chem_water"), ratios = c(.4, .6)))
 
-  blend2 <- purrr::pluck(water2, 5, 1)
+  blend2 <- purrr::pluck(water2, "blended_water", 1)
 
   water3 <- suppressWarnings(water_df %>%
     slice(1) %>%
@@ -208,7 +236,23 @@ test_that("blend_waters_chain can handle different ways to input ratios", {
     ) %>%
     blend_waters_chain(waters = c("balanced_water", "dosed_chem_water"), ratios = c("ratio1", "ratio2")))
 
-  blend3 <- purrr::pluck(water3, 7, 1)
+  blend3 <- purrr::pluck(water3, "blended_water", 1)
 
   expect_equal(blend2, blend3) # test different ways to input ratios
+})
+
+test_that("blend_waters_chain can handle water columns mixed with objects", {
+
+  water4 <- water_df %>%
+    slice(1:3) %>%
+    define_water_chain("A")
+  water5 <- define_water(ph = 7.9, temp = 20, alk = 50, tot_hard = 50, ca = 13, mg = 4, na = 20, k = 20,
+                         cl = 30, so4 = 20, tds = 200, cond = 100, toc = 2, doc = 1.8, uv254 = 0.05)
+
+  blend4 <- water4 %>%
+    blend_waters_chain(c("A", water5), c(.5,.5))
+
+  final <- purrr::pluck(blend4, "blended_water", 1)
+
+  expect_s4_class(final, "water")
 })