+++ tests/testthat/test-chemdose_dbp.R
@@ -27,6 +27,15 @@ test_that("chemdose_dbp warns when inputs are out of model range", {
   expect_warning(chemdose_dbp(water4, cl2 = 4, time = 8)) # br out of bounds
 })
 
+test_that("chemdose_dbp warns about chloramines", {
+  water1 <- suppressWarnings(define_water(ph = 7.5, temp = 20, toc = 3.5, uv254 = 0.1, br = 50, tot_nh3 = 3))
+  water2 <- suppressWarnings(define_water(ph = 7.5, temp = 20, alk = 30, toc = 2, uv254 = 0.01, br = 30)) %>%
+    chemdose_ph(nh42so4 = 3)
+
+  expect_warning(chemdose_dbp(water1, cl2 = 2, time = 8), "breakpoint+")
+  expect_warning(chemdose_dbp(water2, cl2 = 4, time = 8), "breakpoint+")
+})
+
 test_that("chemdose_dbp stops working when inputs are missing", {
   water1 <- suppressWarnings(define_water(toc = 3.5, uv254 = 0.1, br = 50))
   water2 <- suppressWarnings(define_water(ph = 7.5, uv254 = 0.1, br = 5))
@@ -189,7 +198,7 @@ test_that("chemdose_dbp_chain output is list of water class objects, and can han
     balance_ions_chain() %>%
     chemdose_dbp_chain(input_water = "balanced_water", time = 8, cl2 = 4))
 
-  water2 <- purrr::pluck(water1, 8, 1)
+  water2 <- purrr::pluck(water1, "disinfected_water", 1)
 
   water3 <- suppressWarnings(water_df %>%
     mutate(br = 60) %>%
@@ -202,7 +211,7 @@ test_that("chemdose_dbp_chain output is list of water class objects, and can han
     chemdose_dbp_chain(output_water = "diff_name"))
 
   expect_s4_class(water2, "water") # check class
-  expect_equal(names(water3[8]), "diff_name") # check if output_water arg works
+  expect_true(exists("diff_name", water3)) # check if output_water arg works
 })
 
 # Check chemdose_dbp_chain can use a column or function argument for chemical dose